[query] fix read from undefined memory (#12803)
CHANGELOG: Raise an error instead of silently returning bad results when
using a negative index with an `hl.nd.array`.

In practice, I always saw this return a positive value very close to
zero. We should fix this to just support Python-style indexing, but I
wanted to fix ASAP given the severity.