[docker] Trim down service images (#12578)
This is a total overhaul of our docker images. Though very verbose, I
tried to stick to these main tenets:

- Any docker image has exactly 1 layer in it (all the way down to
ubuntu) that installs pip dependencies. This primarily aims to protect
the cache for this particularly large layer and also avoids a later
layer silently upgrading the version of a dependency installed in an
earlier layer. This pairs nicely with the following goal
- We only ever use 1 version of a dependency across the monorepo.
Liberal use of pip's [constraint
files](https://pip.pypa.io/en/stable/user_guide/#constraints-files) to
ensure that the dependencies for a service must be compatible with
dependencies from hail. The `install-dev-dependencies` target which
install all our pinned requirements files would tell you if there's any
incompatible versions of transitive dependencies across the repo
- The image graph is shallow and images don't contain more than they
need. In order to have a single layer with requirements and hail code on
top, I moved the service images to just be based on hail-ubuntu. This
shortens the critical path and therefore reduces total image building
time by reducing the number of times our image data needs to be
downloaded and re-uploaded to the registry. I also removed a lot of
unnecessary cruft like gcloud in places it wasn't used anymore, some
unused/unnecessary pip requirements, etc.