[auth] Make HailCredentials for refreshable auth tokens (#12642)
My motivation here is that in Terra we aren't going to be using hail
authentication tokens, rather an authentication from gcloud or az. So I
want our batch client to be able to use a `CloudCredentials` just as
easily as it uses the hail auth token that we store on the user's
machine. So I introduce HailCredentials which subclasses
CloudCredentials but the behavior is the same. It also lets us re-use
some of the retry logic that we have baked into
`aiocloud.common.Session`.