[compiler] fix overly strict CSE bug (#12819)
The deleted assertion says that a node can't both be lifted to a let,
and be the destination of lets being lifted from below. When I added it,
I figured any lets being lifted from below should be able to lift all
the way to the destination of this node, and I hadn't thought carefully
through what would happen if that assumption weren't true.

The example that breaks that assumption is when some expression `x` is
used as the top level of a branch of an if, or any other spot that
prevents lets from being lifted any higher, while also being used >1
times elsewhere, and `x` also contains a repeated subexpression. The
other uses of `x` will be lifted out somewhere above, so `x` gets marked
as being lifted. The use in the if cannot be lifted, and moreover is the
destination of the lifted repeated subexpression, so `x` gets marked as
being a destination. It's still true that the uses being lifted are
disjoint from the uses which are destinations, but there is only one
object in memory.

I deleted the assertion, added a test case that would have violated it,
and convinced myself that this case is actually handled correctly in
general.