[terraform] align gcp terraform with current state & create new terraform project to track true current state
I fixed some issues in `gcp` but there are enough difficult to resolve issues that I want to start a
fresh terraform project to track reality as different from what we recommend a new user do.

The fresh attempt project is `gcp-broad`. I plan to slowly expand that until it captures everything.
From this point forward, for the subset of infrastructure in `gcp-broad`, we should not make manual
changes. Instead, we should propose changes as PRs, review them, merge them, and apply the terraform
by hand once the changes are in `main`.