[docker] Put Makefile image-building in the top-level Makefile (#12841)
Got annoyed with the constant re-tagging of images that don't need to be
rebuilt, and decided to play a little make golf along the way.

cc @jigold This should dramatically reduce the number of tags for
hail-ubuntu from make-deployed images, though the number of layers in
the container registry should not change.