[qob] retry `storage.writer` and `storage.reader`
I do not think we frequently get errors in `storage.reader`, but I think `storage.writer` was
always flaky and we were protected by the `retryTransientErrors` on `createNoCompression`. My
change to fix requester pays delayed the error until either the first `write` or the `close`
which do not have a `retryTransientErrors` (and it is not obvious to me that it is safe to retry
a `flush`).