[qob] Fix new transiencies (#12869)
Merge of https://github.com/hail-is/hail/pull/12868 and
https://github.com/hail-is/hail/pull/12867 because I expect each is
likely to fail due to the other's error.

---

    [qob] retry `storage.writer` and `storage.reader`
    
I do not think we frequently get errors in `storage.reader`, but I think
`storage.writer` was
always flaky and we were protected by the `retryTransientErrors` on
`createNoCompression`. My
change to fix requester pays delayed the error until either the first
`write` or the `close`
which do not have a `retryTransientErrors` (and it is not obvious to me
that it is safe to retry
    a `flush`).

---

    [qob] retry transient errors reading the results file

---