[qob] Support loading RGs from FASTA files (#12736)
I left the changes to Query and Batch in separate commits for ease of
review. I put these in the same PR because we don't really have
standalone testing for JVM Jobs outside of Query-on-Batch so the FASTA
use-case serves as a test here that cloudfuse is working properly for
JVM Jobs. Would be great if Jackie you could review the batch commit and
Tim could review the query commit.

## Hail Query
- Added support for the `FROM_FASTA_FILE` rpc and the service backend
now passes sequence file information from RGs in every rpc
- Refactored the liftover handling in service_backend to not redundantly
store liftover maps and just take them from the ReferenceGenome objects
like I did for sequence files. This means that add/remove
liftover/sequence functions on the Backend are just intended to sync up
the backend with python, which is a no-op for the service backend.
- Don't localize the index file on fromFASTAFile/addSequence before
creating the index object. `FastaSequenceIndex` just loads the whole
file on construction so might as well stream it in from whatever storage
it's in.
- FASTA caching is left alone because those files will be mounted and
unmounted from the jvm container over the life of the job. JVM doesn't
have to worry about disk usage because that's handled by Batch XFS
quotas, so long as the service backend requests enough storage to fit
the FASTA file. Batch will make sure that a given bucket (and therefore
a given FASTA file) is mounted once per-user on a batch worker.


## Hail Batch
- Added support for read-only cloudfuse mounts for JVM jobs
- These mounts are shared between jobs on the same machine from the same
user
- I did not change DockerJobs, but they could be very easily adapted to
use this new mount-sharing code.