[batch] Setup all of the sql migration infra for deduping resource ids (#12757)
Stacked on #12786 

## Overview:
- The goal of this PR is to setup all of the SQL tables and triggers
such that the next PR can iterate over the attempt resources table to
copy over the correct deduped resource id as well as copy over all of
the 4 billing tables (bp_users, bps, batches, jobs) to collapse rows
with the same deduped resource id.
- I'll put up the subsequent PR as well after this so you can get an
idea for why some of the functionality I added in this PR is there.

## Attempt Resources Table:
- Currently, the attempt resources table has a trigger after insert that
creates entries in the billing tables if the msec_diff != 0.
- We will have merged the mitigation such that for all new attempt
resources entries, resource_id == deduped_resource_id. Therefore, we can
add a before insert trigger that sets deduped_resource_id to
resource_id.
- We do **not** need a column `migrated` here because the only
transformation we want to do with this table is set the
deduped_resource_id to the correct value from the dedup resource id
mapping. We cannot use the attempts or attempt resources table to
populate the new billing tables because we have lost the intermediate
update times and only have time_start and time_ended for each attempt.
- The next migration will populate the correct resource id in 100 row
chunks as we cannot do one massive update (will take days and the
transaction log will be enormous).
- The `attempt_resources_after_insert` table needs to only populate the
new `v3` billing tables if the rows for that primary key have already
been copied from the `v2` tables (i.e. migrated=1 on the row in the `v2`
billing table). The issue here is that we need to account for the time
in which we have not copied a row from the v2 billing table, but we are
already adding it to the `v3` table. This would mean we'd end up double
counting:

For example,

Say we have this row currently in the agg_bp_resources_v2 table:
(test_bp, compute, 100, migrated=0)

Say we insert this row into the attempt_resources table:
(1, 1, bp=test_bp, resource=compute, 50)

The agg_bp_resources_v2 table will be
(test_bp, compute, 150, migrated=0)

If we add the usage to the `v3` table (not what I am proposing to do):
(test_bp, compute, 50)

Then later on when we copy over rows with migrated=0 in the `v2` billing
tables, we'll end up with this row in the `v3` table because we need a
`+=` here to account for newly added records:
(test_bp, compute, 200)

It should be (test_bp, compute, 150).

Therefore, for the `attempt_resources_after_insert` trigger, we need to
check whether we have migrated the row already (migrated = 1 on the
corresponding billing table) before inserting / updating the billing
table rows in `v3`.

## Attempts Table
- There's a trigger `attempts_after_update` here that updates the usage
of all four `v2` billing tables based on the start and rollup times.
- Like for the attempt_resources table, we only want to add/update rows
into `v3` billing tables if the corresponding rows for `v2` have already
been migrated.
- This is where the extra join comes in as a filter:
```
    JOIN aggregated_job_resources_v2 ON
      aggregated_job_resources_v2.batch_id = attempt_resources.batch_id AND
      aggregated_job_resources_v2.job_id = attempt_resources.job_id AND
      aggregated_job_resources_v2.resource_id = attempt_resources.resource_id AND
      aggregated_job_resources_v2.token = rand_token
    WHERE batch_id = NEW.batch_id AND job_id = NEW.job_id AND attempt_id = NEW.attempt_id AND migrated = 1
```

## Aggregated Billing Tables v2
- We need to add a before insert trigger to set migrated=1 for all new
rows that get inserted in each of the corresponding tables
- We need after update triggers to copy the rows from the v2 table into
the corresponding v3 table if OLD.migrated = 0 and NEW.migrated = 1.

## Concerns
- The only qualm I have right now with the design I've described above
is making sure we can change the attempt_resources table
`deduped_resource_id` into the primary key of the table and eventually
remove the current `resource_id` without it taking forever and blocking
the table. I'll need a clone of the database to determine this.