[fs] Support reading from public containers in azure (#12900)
CHANGELOG: Reading data from public blobs is now supported in Azure

On current main, the following test fails:
```python
def test_can_access_public_blobs():
    public_mt = 'hail-az://azureopendatastorage/gnomad/release/3.1/mt/genomes/gnomad.genomes.v3.1.hgdp_1kg_subset.mt'
    assert hl.hadoop_exists(public_mt)
    mt = hl.read_matrix_table(public_mt)
    mt.describe()
```

The `hadoop_exists` fails in the python fs with a
`ClientAuthenticationError` and the QoB pipeline fails on the Query
Driver with a 401. See
[zulip](https://hail.zulipchat.com/#narrow/stream/300487-Hail-Batch-Dev/topic/Azure.20public.20blobs/near/284186815)
for additional context. I contemplated whether we should have a public
container in our infrastructure for running tests, but the lack of
requester pays makes that feel not great. Azure's paying for the
opendatastorage account so I figured it would be ok for us to test
against it (and the test only reads metadata), but happy to take
alternative suggestions.