[docs, minor] Prefer missing() rather than deprecated null() (#12902)
https://hail.is/docs/0.2/overview/expressions.html#missingness says:

> In Hail, all expressions can be missing. An expression representing a
missing value of a given type can be generated with the
[`null()`](https://hail.is/docs/0.2/functions/core.html#hail.expr.functions.null)
function, which takes the type as its single argument.

It then proceeds to show a number of examples, all of which use
`missing()` rather than `null()`, which is confusing.

PR #9936 added `hl.missing`, which supersedes `hl.null`, and replaced
instances of `null(...)` invocations in the documentation. However
`:func:'.null'` instances in the surrounding documentation text were
missed.