[hailtop] Add hailtop.fs for user-level RouterFS functions (#12731)
CHANGELOG: Introduce `hailtop.fs` that makes public a filesystem module
that works for local fs, gs, s3 and abs. This is now used as the
`Backend.fs` for hail query but can be used standalone for Hail Batch
users by `import hailtop.fs as hfs`.

Still have to do the docs but a couple questions remain:

I create a hidden singleton `RouterFS` object so that is used by
functions in `hailtop.fs`. Should this singleton also be used by the
Hail Query backends when they are initialized? How do we propagate
configuration information such as `requester_pays_bucket` to the FS?

---------

Co-authored-by: Dan King <dking@broadinstitute.org>