[is_transient_error] ensure we recur for retry_once and clean up is_t… (#12930)
…ransient_error a bit

I saw another `ConnectionResetError` that [was not
retried](https://ci.hail.is/batches/7309548/jobs/105). It was hidden as
the cause of a `ClientOSError`. This change ensures that both
`is_retry_once_error` and `is_transient_error` both recur through the
cause hierarchy regardless of the error type.

I also unified the formatting of `is_transient_error` conditions and
ensured that we fall through the `if`s in every case except docker. The
docker error case was complex and, AFAIK, docker does not raise errors
with causes that matter.