[batch] Add timeouts to ensure that post_job_complete is reached (#12944)
We always ensure that `cleanup` is run, but in order to ensure that
`post_job_complete` is run we need to ensure we get through any
computation in cleanup and `mark_complete` that could potentially hang.
So we add timeouts to any yield points in those functions and broadly
catch exceptions so we always continue in the cleanup/mark_complete
process regardless of failure.