[query] fix & improve pprint for hl.Struct (#12901)
CHANGELOG: `hl.Struct` now has a correct and useful implementation of
pprint.

For structs with keys that were not identifiers, we produced incorrect
`repr` output. For `pprint`, we just `pprint`'ed a dictionary (so you
cannot even tell that the object was an `hl.Struct`). This PR:

1. Fixes `hl.Struct.__str__` to use the kwargs or dictionary
representation based on whether the keys are Python identifiers.
2. Teaches `StructPrettyPrinter` to first try to `repr` the struct (this
is what the default pretty printer does)
3. Teaches `StructPrettyPrinter` to properly pretty print a struct as an
`hl.Struct` preferring the kwarg representation when appropriate.
4. Teaches `_same` to use pretty printing when showing differing
records.