[batch] use regional bucket for requester pays tests (#12964)
I updated terraform but

1. GCP Terraform state is still local on my laptop.

2. GCP Terraform appears to not configure global-config. As such, I
cannot thread the name of the bucket through to the tests the way we do
with TEST_STORAGE_URI. For now, I've hardcoded the name (which is what
we were doing previously). When we eventually get to testing recreation
of GCP in a new project we'll have to address the global config then.