[services] indicate how many errors we have seen
We are increasingly seeing errors from "Connection reset" which we switched from
"transient" to "retry once". The current code makes it impossible to determine
if we are correctly retrying this error once.

If we see that there are a lot of "Connection reset" errors that happen twice
we should perhaps change "retry once" to "retry five times" or use a more generous
delay.