[hailctl/dataproc] upgrades `notebook` to version 6.0.* (#12965)
Attempting to install hail on dataproc using the `init_notebook.py`
script errors out because hail depends on a more recent version of
`jinja2` than the version of `notebook` pinned in the script allows for.
This change upgrades the pinned version of `notebook` in the script to
be compatible with hail's pinned `jinja2` version.

Fixes #12926.