[batch] Delete old unused docker socket field in job spec (#12973)
This has been unused for a long time and I wanted to rid the world of
it. This changes the client to no longer send the `mount_docker_socket`
field, updates the validator so it's ok for that field to be missing,
and explicitly rejects any requests where `mount_docker_socket` is set
to true. The reason I couldn't remove `mount_docker_socket` entirely
from the codebase is older clients would break if it was not at least
optional in the validator spec, which is kind of annoying.

Dev deployed this and successfully submitted jobs using the client from
this branch and the client from main.