[qob] retry GSFS reads
After an `FSSeekableInputStream` method (successfully) returns, `getPosition` always represents the
intended location within the object. We can entirely eliminate `lazyPosition` because it tracks the
same value as `getPosition` when `reader == null`.

For retryable reads, we just seek back to the known correct location of the stream before attempting
to read again.