[batch] Simplify cloudfuse credential handling (#12962)
Just a refactor, simplifies the interactions between `Worker`,
`CloudWorkerAPI` and `CloudUserCredentials` and hopefully makes this
code safer and easier to work with. Instead of the following occuring in
worker.py:

1. get credentials string from `CloudUserCredentials`
2. tell `CloudWorkerAPI` to write credentials string to `path` owned by
the job
3. tell `CloudWorkerAPI` to mount cloudfuse using the credentials stored
at `path`

we instead just do

1. tell `CloudWorkerAPI` to mount cloudfuse using `CloudUserCredentials`

On its own I think this change makes the codepath simpler and easier to
think about in terms of where credentials are stored, but this also gets
rid of the requirement from `worker.py`'s point of view that credentials
must be stored on the filesystem. This will make it easier to transition
off of key files and over to metadata server tokens.

In order to make the new statement sound in terms of types, we can't
have `CloudWorkerAPI.mount_cloudfuse` just accept a
`CloudUserCredentials` argument, because that means `GCPWorkerAPI` would
need to be able to support an argument of type `AzureUserCredentials`,
which would never happen and doesn't make sense. What we can do here is
make `CloudWorkerAPI` generic over the subtype of `CloudUserCredentials`
that it both produces and consumes. This allows us to use stricter types
like `GCPUserCredentials` and `AzureUserCredentials` inside of
`GCPWorkerAPI` and `AzureWorkerAPI` respectively and now the type system
is happy. It also relaxes the restriction that both of the
`GCPUserCredentials` and `AzureUserCredentials` need to conform to the
same `cloudfuse_credentials` interface.