[ci] Add CreateDatabase2Step for in-namespace test and dev databases (#13022)
These changes are a prerequisite for introducing a mysql DB pod for
every test and dev namespace. The crux of such a change is any
CreateDatabase steps should use the `database-server-config` from *its
own namespace* (which will come in the PR that uses this step) and not
from default. There's no cleanup step required because this will be used
to create DBs inside the namespace for the pipeline, so resources will
get cleaned up with the namespace. The other changes in this step bring
the configuration for `dev` scopes closer to that of `test` scopes,
because creation of test databases should really just be idempotent and
there shouldn't be a difference between deploying a database in dev and
test. I would have deferred making the changes to the `dev` scope except
dev deploy was the most practical way for me to test this change.