[ci] Create a test MySQL server in test and dev namespaces (#13030)
This change creates mysql pods for test and dev namespaces, instead of
sharing the CloudSQL database server. The areas of change are as
follows:

### Generation of the namespace's database-server-config
The current approach in main does a little trick. Since the current
`createDatabase` step uses the `database-server-config` from default to
generate admin/user sql configs, the CI pipeline creates a dummy
database `test-database-instance` to create a
`sql-test-instance-admin-config` that inherits the credentials from the
production `database-server-config`, and then copies that within the
test namespace to `database-server-config`.

In this change, since we are creating the server ourselves, we can just
replace these with a step that creates a `database-server-config` from
scratch, and then uses that for the DB pod.

Overall making these changes really gave me the heebie jeebies that the
test and dev namespaces have all these credentials to the CloudSQL
server. I'm glad this gets rid of that.

### Accessing the database server
We use the DB pod's service DNS name as the `host` so inside Kubernetes
this Just Works. The one caveat is the CI pipeline in which we run
migrations in batch jobs. Those jobs need a way to reach the DB pod. I
achieve this with a NodePort and then use the job's K8s credentials to
resolve the node and port that the DB is on. The code I've added to do
this resolution feels a bit janky, wouldn't mind some feedback on that.
In terms of security, if a user job was able to somehow resolve the
address of a test db, they would still not have the credentials to
access it, and this is currently also the case with the production
database. Nevertheless, this does raise an action item that we should
only allow traffic to the k8s and DB subnets for `network=private` jobs,
but I think we should make that a separate PR.

### Database creation
In order to test this properly in a dev deploy, I needed to make some
changes to `create_database.py`. In main, dev deploys can't create
databases. I think they should be able to, and those operations should
just be idempotent. When allowing dev deploys to create databases, I hit
the `ALTER USER` lines in `create_database.py` which lock out any
already-deployed application, which feels silly. Instead, I create the
mysql user and create the corresponding secret iff the mysql username
does not already exist.

### create_initial_account.py
When starting off with a fresh database, we encounter a bootstrapping
problem where we need to add the developer's user. The current way to
add the developer's user is through `create_initial_account.py` which
assumes that the developer's gsa key secret is not already in the
namespace, but now it could be, so I delete the key in the namespace if
it exists before creating it. This could all change with my other PR to
add devs to test namespaces. But this change to allow ephemeral dev
databases will make testing that other PR *way* easier.