[query] Expand ReadValue beyond hail EType deserialization (#12948)
This represents a small redesign of ValueWriter as well.

Add deserializers, subclasses of ValueReader and have ReadValue use
them. A ValueReader deserializes a single hail value from an input
stream.

This change also alters the semantics of ValueWriters. Both readers and
writers no longer manage their own I/O resources. It is the
responsibility of 'callers' to do so instead. However programmers must
be careful as we need to create input/output buffers to perform native
serialization/deserialization. For InputBuffers in particular, the
underlying stream may be left in an unusable state.