[batch] Use same size VMs in dev and PR as we use in prod (#12974)
Now that test databases are hosted on their own servers instead of the
single cloud-hosted MySQL, we can ramp up the parallelism both in our
tests and in the number of PRs that we run at once. I recall that even
before we had this DB bottleneck we still restricted the number of PRs
running at once for cost reasons, but if that's not the case we could
remove that restriction entirely.