CHANGELOG Allow subnet to be passed through to gcloud in hailctl (#13129)
`hailctl` allows the user to define a `network` for a new dataproc
cluster, but not a `subnet`. The `network` flag in `gcloud` only works
for `auto` and `legacy` project networks. If a Google Project has a
custom network, the `subnet` must be supplied manually.

This PR adds that functionality to `hailctl`. Its a very simple
passthrough, so not much effort was required.