[qob] tightly scope retries in ServiceBackend (#13152)
We get a lot of errors about files that already exist. Hail commands are
usually not retryable because there are file paths that might have been
partly written to. This tightly scopes the retries to just the I/O. It
also avoids keeping the output stream open for a long period of time.