[fs] Fix globbing in the scala blob storage FS (#13173)
Using URL parsing for object storage URIs has proven problematic as they
don't really have the same syntax. In particular, Java URI truncates the
path at `?` and assigns the rest to a query string field (as it should),
but `?` is a valid character in a blob storage object and is used as a
glob pattern. We were forgetting to look at the `URI.query` field to
assemble the full object name and therefore broke globbing. I added a
few tests specifically that used single-character match pattern (`?`)
and got rid of `URI` in favor of a regex to get ahead of any other
potential footguns.