[hailctl] Replace argparse with typer/click (#13109)
Started playing with [typer](https://typer.tiangolo.com/) on the plane.
It's a somewhat thin wrapper around `click`, and I'm not entirely sold
on one vs the other, but either seems a lot neater than argparse (I
remember Cotton tried this years ago but I haven't looked at that PR).
Click gives us the easy composition of CLIs, function name as the
command name, the use of docstring as the description, help output, and
most of the core functionality you see. Typer is the one translating
python type hints into click type-checking, default setting, etc. Typer
also uses rich if it's installed and allows you to install shell
completion which is pretty neat, not sure if you can get that through
click or if you need to use a click plugin.

I replaced a couple of the hailctl commands here with the typer/click
version. If you like what you see I can keep going on the rest of the
commands (you can look to `batch/cli.py` and `hailctl/__main__.py` as an
example).